Update get_eol function
The get_eol function was refactored for efficiency and clarity in the hymetDP
package. These changes, including new unit test and updated instances, are now
ported over to ecocomDP.

Add test for the new get_eol function

This test was developed in the hymetDP package but since the function and test
should be essentially the same, it is a direct copy.

Remove unused argument from get_eol instance