Fix test warnings
Half a dozen tests (or so) were throwing warnings related to deprecated
usage of external vector references in selections.

This commit applys the recommended tidyselect fixes. Except for one case
where the dplyr::relocate() is being called, and for which a solution
has not yet been found. This last test is for the flatten_data()
function.