Fix vignette index warnings
During the source package build and check process, a series of warnings
were issued related to the vignette title specified in the vignette
index entry being different from the title in the YAML metadata.

This commit fixes these issues by listing the vignette title, in the
header block, in the fignette index entry.