Fix build notes
A note occuring in the build process cautioned against comparing a
class() call to a string, which in the case of multi-class objects,
would return a vector of boolean values rather than a single boolean
value. This predisposes the code to bugs.

This commit fixes this logic by checking that the string is in the
results of the class() call rather than the other way around.