Clarify a couple features of search_data()
* Clarify logic for combining search parameters

Clarify the logic used for combining multiple search parameters to
address confusion by some users on observed result sets. State the
implicit "AND" boolean logic to reframe users expectations.

Consider adding user control over this behavior through the
addition of an new parameter, or adjustment of the existing
"boolean" parameter.

* Clarify the taxa parameter of search_data()

Rather than explain nuances of the implementation and leave it up to the
user to translate them into how the parameter should be utilized, provide
the user a simple process to follow to optimally search the taxonomic
tree.