Implement traits extension into R package code
Implement the traits extension into the R package, with the exception of
renaming the `variable_mapping` table to `variable_attribute` to
preserve backwards compatibility while this issue is under discussion.

Note, this doesn't touch the NEON mapping functions. These will need to
be updated in a subsequent commit.

Related to #158 and #167